Home page refactored - added new component - HomeCards - to display cards in home page
